+++ src/FluentlyHttpClient/Utils/QueryStringOptions.cs
@@ -48,6 +48,7 @@ public record QueryStringOptions
 	/// Gets or sets the function to format a collection item. This will allow you to manipulate the value.
 	/// </summary>
 	public Func<object, string>? CollectionItemFormatter { get; set; }
+	public Dictionary<string, Func<object, string>>? CollectionValuePerKeyItemFormatter { get; set; }
 
 	/// <summary>
 	/// Gets or sets the function to format the key e.g. lowercase.
@@ -55,6 +56,7 @@ public record QueryStringOptions
 	internal Func<string, string> KeyFormatter { get; set; } = DefaultKeyFormatter;
 
 	internal Func<object, string>? ValueFormatter { get; set; }
+	internal Dictionary<string, Func<object, string>>? ValuePerKeyFormatter { get; set; }
 
 	internal Func<string, string> ValueEncoder { get; set; } = DefaultValueEncoder;
 
@@ -95,6 +97,18 @@ public record QueryStringOptions
 	{
 		ValueFormatter = configure;
 		CollectionItemFormatter = configure;
+		return this;
+	}
+
+	/// <summary>
+	/// Gets or sets the function to format a value per key. This will allow you to manipulate the value by key
+	/// </summary>
+	/// <param name="configure"></param>
+	/// <returns></returns>
+	public QueryStringOptions WithValuePerKeyFormatter(Dictionary<string, Func<object, string>> configure)
+	{
+		ValuePerKeyFormatter = configure;
+		CollectionValuePerKeyItemFormatter = configure;
 		return this;
 	}